requirements.txt
aiohttp==3.8.1
disnake==2.8.0
python-dotenv==1.0.0
